v5第三次提交

+++ yudao-module-steam/yudao-module-steam-biz/src/main/java/cn/iocoder/yudao/module/steam/service/fin/v5/res/V5ProductPriceInfoRes.java
@@ -31,10 +31,10 @@ public class V5ProductPriceInfoRes {
             private String itemQuality;
             private String itemQualityName;
             private Integer onSaleStock;
-            private Integer minSellPrice;
+            private Double minSellPrice;
             private String itemIcon;
-            private Integer selfPrice;
-            private Integer playerPrice;
+            private Double selfPrice;
+            private Double playerPrice;
 
         }
     }
